# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sigilum API",
    "description": "Agent identity and authorization registry. Manages namespaces, claims, verification, DID resolution, webhooks, and dashboard authentication.",
    "version": "0.1.0",
    "contact": {
      "name": "Sigilum",
      "url": "https://sigilum.id"
    }
  },
  "servers": [
    {
      "url": "https://api.sigilum.id",
      "description": "Production"
    },
    {
      "url": "http://localhost:8787",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/v1/namespaces/claims": {
      "get": {
        "operationId": "listApprovedServiceClaims",
        "tags": [
          "Namespaces"
        ],
        "summary": "List approved claims for authenticated service",
        "description": "Returns approved claims for the service tied to the provided API key.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "service",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Must match the authenticated service slug when provided"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 2000,
              "default": 500
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Approved claims",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovedNamespaceClaimsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/claims": {
      "post": {
        "operationId": "submitClaim",
        "tags": [
          "Claims"
        ],
        "summary": "Submit claim",
        "description": "Submit an authorization request on behalf of an agent. Requires a service API key.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitClaimRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Existing pending or approved claim",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitClaimResult"
                }
              }
            }
          },
          "201": {
            "description": "Claim created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitClaimResult"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/verify": {
      "get": {
        "operationId": "verifyAgent",
        "tags": [
          "Verify"
        ],
        "summary": "Verify agent authorization",
        "description": "Check whether a public key is currently approved for a service/namespace pair.",
        "parameters": [
          {
            "name": "namespace",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "public_key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "service",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Verification result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/services": {
      "get": {
        "operationId": "listServices",
        "tags": [
          "Services"
        ],
        "summary": "List services",
        "security": [
          {
            "CookieAuth": []
          },
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Services",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createService",
        "tags": [
          "Services"
        ],
        "summary": "Create service",
        "security": [
          {
            "CookieAuth": []
          },
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServiceRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Service created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateServiceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Service API key (sk_live_...) or dashboard JWT"
      },
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sigilum_token",
        "description": "Dashboard auth cookie"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "details": {}
        }
      },
      "Pagination": {
        "type": "object",
        "required": [
          "limit",
          "offset",
          "total",
          "has_more"
        ],
        "properties": {
          "limit": {
            "type": "integer",
            "minimum": 0
          },
          "offset": {
            "type": "integer",
            "minimum": 0
          },
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "has_more": {
            "type": "boolean"
          }
        }
      },
      "ClaimStatus": {
        "type": "string",
        "enum": [
          "pending",
          "approved",
          "revoked",
          "rejected",
          "expired"
        ]
      },
...
      "ApprovedNamespaceClaim": {
        "type": "object",
        "required": [
          "claim_id",
          "namespace",
          "public_key",
          "service",
          "approved_at"
        ],
        "properties": {
          "claim_id": {
            "type": "string"
          },
          "namespace": {
            "type": "string"
          },
          "public_key": {
            "type": "string"
          },
          "service": {
            "type": "string"
          },
          "approved_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ApprovedNamespaceClaimsResponse": {
        "type": "object",
        "required": [
          "claims",
          "pagination"
        ],
        "properties": {
          "claims": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApprovedNamespaceClaim"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        }
      },
      "VerifyResponse": {
        "type": "object",
        "required": [
          "authorized"
        ],
        "properties": {
          "authorized": {
            "type": "boolean"
          },
          "claim_id": {
            "type": "string"
          },
          "approved_at": {
            "type": "string",
            "format": "date-time"
          },
          "reason": {
            "type": "string"
          }
        }
      },
      "SubmitClaimRequest": {
        "type": "object",
        "required": [
          "namespace",
          "public_key",
          "service",
          "agent_ip",
          "nonce"
        ],
        "properties": {
          "namespace": {
            "type": "string",
            "minLength": 3,
            "maxLength": 64
          },
          "public_key": {
            "type": "string"
          },
          "service": {
            "type": "string",
            "minLength": 1
          },
          "agent_ip": {
            "type": "string"
          },
          "nonce": {
            "type": "string",
            "minLength": 8,
            "maxLength": 256
          }
        }
      },
      "SubmitClaimResult": {
        "type": "object",
        "required": [
          "claim_id",
          "status",
          "service",
          "message"
        ],
        "properties": {
          "claim_id": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ClaimStatus"
          },
          "service": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "ServiceWebhookBootstrap": {
        "type": "object",
        "required": [
          "url",
          "secret"
        ],
        "properties": {
          "url": {
            "type": "string",
            "format": "uri",
            "maxLength": 2048
          },
          "secret": {
            "type": "string",
            "minLength": 16
          },
          "auth_header": {
            "type": "string",
            "maxLength": 128
          },
          "auth_value": {
            "type": "string",
            "maxLength": 2048
          }
        }
      },
      "CreateServiceRequest": {
        "type": "object",
        "required": [
          "name",
          "slug",
          "domain",
          "description"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "slug": {
            "type": "string",
            "minLength": 3,
            "maxLength": 64,
            "pattern": "^[a-z0-9][a-z0-9-]{1,62}[a-z0-9]$"
          },
          "domain": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          },
          "webhook": {
            "$ref": "#/components/schemas/ServiceWebhookBootstrap"
          }
        }
      },
      "ServiceSummary": {
        "type": "object",
        "required": [
          "id",
          "name",
          "slug",
          "domain",
          "description",
          "created_at",
          "updated_at",
          "registration_tx_hash",
          "active_key_count"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "domain": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "registration_tx_hash": {
            "type": [
              "string",
              "null"
            ]
          },
          "active_key_count": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "ServiceListResponse": {
        "type": "object",
        "required": [
          "services"
        ],
        "properties": {
          "services": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServiceSummary"
            }
          }
        }
      },
      "CreateServiceResponse": {
        "type": "object",
        "required": [
          "id",
          "name",
          "slug",
          "domain",
          "description",
          "created_at",
          "registration_tx_hash"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "domain": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "registration_tx_hash": {
            "type": [
              "string",
              "null"
            ]
          },
          "webhook": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "url": {
                "type": "string",
                "format": "uri"
              },
              "events": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "required": [
              "id",
              "url",
              "events"
            ]
          }
        }
      },
...
    }
  }
}